+++ src/queuebindframe.h
@@ -87,7 +87,7 @@ public:
      *  @param   Table arguments     additional arguments
      */
     QueueBindFrame(uint16_t channel, const std::string& name, const std::string& exchange, const std::string& routingKey = "", bool noWait = false, const Table& arguments = {}) :
-        QueueFrame(channel, (name.length() + exchange.length() + routingKey.length() + arguments.size() + 6) ), // 3 extra per string, 1 for bools, 2 for deprecated field
+        QueueFrame(channel, (uint32_t) (name.length() + exchange.length() + routingKey.length() + arguments.size() + 6) ), // 3 extra per string, 1 for bools, 2 for deprecated field
         _name(name),
         _exchange(exchange),
         _routingKey(routingKey),
